+++ CHANGELOG.md
@@ -7,13 +7,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.10.0] - 2023-10-23
+
+## Added
+
+- Added heat pump state "inflow"
+- Added `heat_circuit.set_holiday_temperature()` endpoint
+
 ## [1.9.0] - 2023-10-23
 
 ## Added
 
 - Add **AWAY** and **PARTY** to heat circuit operating mode
 
-## [1.8.1] - 2023-10-23
+## [1.8.1] - 2023-10-22
 
 ## Added